+++ app/src/main/java/com/example/des/hp/Tip/TipDetailsView.java
@@ -16,6 +16,8 @@ import com.example.des.hp.R;
 
 import static com.example.des.hp.Database.DatabaseAccess.databaseAccess;
 
+import androidx.annotation.NonNull;
+
 public class TipDetailsView extends BaseActivity
 {
 
@@ -39,12 +41,12 @@ public class TipDetailsView extends BaseActivity
             layoutName="activity_tip_details_view";
             setContentView(R.layout.activity_tip_details_view);
 
-            txtTipDescription=(TextView) findViewById(R.id.txtTipDescription);
-            txtTipNotes=(TextView) findViewById(R.id.txtTipNotes);
-            btnClear=(ImageButton) findViewById(R.id.btnClear);
-            btnSave=(Button) findViewById(R.id.btnSave);
-            grpTipDescription=(LinearLayout) findViewById(R.id.grpTipDescription);
-            grpMenuFile=(LinearLayout) findViewById(R.id.grpMenuFile);
+            txtTipDescription= findViewById(R.id.txtTipDescription);
+            txtTipNotes= findViewById(R.id.txtTipNotes);
+            btnClear= findViewById(R.id.btnClear);
+            btnSave= findViewById(R.id.btnSave);
+            grpTipDescription= findViewById(R.id.grpTipDescription);
+            grpMenuFile= findViewById(R.id.grpMenuFile);
 
             afterCreate();
 
@@ -74,21 +76,14 @@ public class TipDetailsView extends BaseActivity
 
     //region OnClick Events
     @Override
-    public boolean onOptionsItemSelected(MenuItem item)
+    public boolean onOptionsItemSelected(@NonNull MenuItem item)
     {
-        try
-        {
-            switch(item.getItemId())
-            {
-                case R.id.action_delete_tip:
-                    deleteTip();
-                    return true;
-                case R.id.action_edit_tip:
-                    editTip();
-                    return true;
-                default:
-                    return super.onOptionsItemSelected(item);
-            }
+        try {
+            int id = item.getItemId();
+            if (id == R.id.action_delete_tip)
+                deleteTip();
+            if (id == R.id.action_edit_tip)
+                editTip();
         }
         catch(Exception e)
         {
@@ -105,14 +100,14 @@ public class TipDetailsView extends BaseActivity
         try
         {
             tipItem=new TipItem();
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 if(!da.getTipItem(holidayId, tipGroupId, tipId, tipItem))
                     return;
             }
 
 
-            if(title == null || (title.length() == 0))
+            if(title == null || (title.isEmpty()))
             {
                 SetTitles(tipItem.tipDescription, "");
             } else
@@ -155,7 +150,7 @@ public class TipDetailsView extends BaseActivity
         try
         {
             tipItem.noteId=pNoteId;
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 da.updateTipItem(tipItem);
             }
@@ -187,7 +182,7 @@ public class TipDetailsView extends BaseActivity
         try
         {
             tipItem.infoId=pInfoId;
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 da.updateTipItem(tipItem);
             }
@@ -224,7 +219,7 @@ public class TipDetailsView extends BaseActivity
     {
         try
         {
-            try(DatabaseAccess da = databaseAccess();)
+            try(DatabaseAccess da = databaseAccess())
             {
                 if(!da.deleteTipItem(tipItem))
                     return;
